correction

+++ 0x00-hello_world/6-size.c
@@ -6,16 +6,16 @@
  */
 int main(void)
 {
-	char a;
-	int b;
-	long c;
-	long long d;
-	float e;
+char a;
+int b;
+long c;
+long long d;
+float e;
 
-	printf("the size of a char is: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("the size of an int is: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("the size of a long is: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("the size of a long long is: %lu byte(s)", (unsigned long)sizeof(d));
-	printf("the size of a float is: %lu byte(s)", (unsigned long)sizeof(e));
-		return(0);
+printf("the size of a char is: %ld byte(s)\n", sizeof(a));
+printf("the size of an int is: %ld byte(s)\n", sizeof(b));
+printf("the size of a long is: %ld byte(s)\n", sizeof(c));
+printf("the size of a long long is: %ld byte(s)", sizeof(d));
+printf("the size of a float is: %ld byte(s)", sizeof(e));
+return(0);
 }
